Kelso drops Iridiums 15 and 56. Sladen drops 15.
This is an unusual situation, because both Kelso and Sladen have dropped
Iridium 15, but Space Track claims it is still on orbit. Normally I
consider Space Track definitive for on-orbit status and Kelso for
operational status, but I think it's too much to re-inject a body that
both Kelso and Sladen have dropped.

Current plan: Drop 15 from Astro::Coord::ECI::TLE::Iridium but not
Astro::SpaceTrack, and wait to see what happens.

Iridium 56 has already been recorded as decayed.